Refactor import resources UI handling
